+++ config/environments/development.rb
@@ -3,7 +3,7 @@ require "active_support/core_ext/integer/time"
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
-  config.force_ssl = true
+  config.force_ssl = false
   config.ssl_options = { hsts: false }
   config.hosts << "heimdall.dev"
   config.hosts << "localhost"