+++ config/environments/production.rb
@@ -29,6 +29,14 @@ Rails.application.configure do
 
   # Force all access to the app over SSL, use Strict-Transport-Security, and use secure cookies.
   config.force_ssl = true
+  config.ssl_options = {
+    redirect: { exclude: ->(req) { req.path == "/up" } }
+  }
+
+  # Trust reverse proxy headers
+  config.action_dispatch.trusted_proxies = [
+    IPAddr.new("127.0.0.1"), IPAddr.new("::1")
+  ]
 
   # Skip http-to-https redirect for the default health check endpoint.
   # config.ssl_options = { redirect: { exclude: ->(request) { request.path == "/up" } } }